fix a bug when playing sound
When I tried to play on VBox, I found that it can't play fully, then I found that the `played_bytes` might be the same as `buffer_size`. This will cause it to do nothing but loop in two branches, and finally the played size will be more than the size_of_full_pcm_output_in_bytes, so it calls `step_sound()`, the card stop playing the pcm data. and in fact, the pcm data isn't played fully.